Merge .appup.pre.src & .appup.post.src files into the generated appup
Add pre and post instructions to the instuctions created by appup generate.
These instructions must be stored in the .appup.pre.src and .appup.post.src
files in the src folders of the given application.

If one of these files are missing or the version patterns specified in
these files don't match the current old and new versions stored in the
.appup file the corresponding part will be empty. If none of these files
are present the generated appup file will be exactly the same as without
using this feature.